refactor(coworker): split briefing loader out of the action module (module-size ratchet)
buildOpeningBriefingPayload moves to lib/agent/opening-briefing-loader.ts;
agent-coworker.ts keeps only the call + snapshot field (+13 lines net,
re-baselined 2638 -> 2651 with justification; union-merge baseline).
Typecheck evidence: tsc --noEmit exit 0 with 8GB heap (hook tsc OOMs cold).
